Open Kafka and Zookeeper ports.
If these ports are not open, no client outside the docker containers can consume
the contents of the statediff topic, preventing the use of the diffPublish
feature.

In order to connect to kafka using these ports, you must alias the machine on
which it's running as 'kafka', and likewise for zookeeper.  For instance, to do
this for your local machine, add 'kafka' and 'zookeeper' to the end of the line
in /etc/hosts starting with '127.0.0.1 localhost'.