simplify everything
```
_swaggerHost :: Maybe Host
The host (name or ip) serving the API. It MAY include a port. If the
host is not included, the host serving the documentation is to be used
(including the port).
_swaggerBasePath :: Maybe FilePath
The base path on which the API is served, which is relative to the
host. If it is not included, the API is served directly under the host.
The value MUST start with a leading slash (/).
```