STRATO-359: Reenable test coverage for strato-api
Files named "*Spec_excluded.hs" are not picked up by hspec-discovery,
as they must end in "Spec.hs". In addition to enabling those tests,
    - withApp was changed from `SpecWith App` to `SpecWith (TestApp App)`
    - wipeDB and runMigration is enabled for all tables. However, this
      does not yet create databases for "peers", which is why PeersSpec
      is disabled
    - Broken urls were changed from e.g. /query/block to using the type
      safe RequestBuilder with e.g. BlockInfoR.

In addition to the peer database creation, this also does not enabled
FromJSON for the Block so most of the interesting tests in JsonSpec.hs
cannot yet work.