Give guidance on how to create matching sol, bin pairs
Trying to use solc --bin-runtime is a tar pit because of the metadata
that bloch adds and its sensitivity to any changes to the codehash.

To that end, the Makefile pretending that method work is removed
and the bin-runtime for the sample contract is checked in.
Test files that I never ended up using are deleted.