Write an initial statediff to kafka for cirrus to pick up.
For this, it was necessary to alter the backend of writeStateDiffs
to limit the message size to kafka. Additionally, kafka can
return an error that is not Left, so it was necessary to filter
the ProduceResponse as well for real errors.

There's optimization to be done in choosing a message size.
Linkedin ran a study that concluded that 10KB will be best
for kafka throughput, but currently even after all publishes
have been made cirrus is making its way through a large backlog
since it needs to GET /storage for every message (or I may just
be waiting for its log to clear a backlog).