Remove caching issue for builds of solc.
Currently its pretty hard to update if our binaries need to change in sync
with solc. If solc is already cloned, the makefile will never change it again
This change forces a clone/checkout if the tag in SOLCBRANCH doesn't match
a current tag on HEAD in the local solidity repo, if it exists.