Convert extraData from Integer to ByteString
Its description in the yellow paper is as a member of \mathbb B,
the class of bytestrings, with the additional caveat of being less
at most 32 bytes. I intend to ignore the part about 32 bytes, but
converting to a byte array easier to work with in pbft.

The GenesisInfo's extraData is not converted to preverve
backwards compatibility with Genesis.json files.

Note as well that since postgres uses varchar for the Integer types,
no migration is needed. Legacy extraDatas will just be ['0'], and
new extradatas will be [].