Archive docker-compose.yml after build in test pipeline
I haven't done the same for the release pipeline, to make
it harder to accidentally release a bad docker-compose.yml.
But a lot of times, I would want a copy of a bad docker-compose.yml
locally so I can try to deploy or run the tests and diagnose what
failed.