Be more generous in accepting proposer seals
From the logs in https://blockappsdev.slack.com/archives/G5E7K3ETX/p1539727000000100,

At the time of block #95, there were 5 nodes out to lunch[0]. During its
initial announcement by 315, 3 nodes become locked on #95: 311, 315, 317
[1]. Quorum in a 20 node network is 14 (note: I had miscounted yesterday
as 13, but 13 * 3 < 2 * 20 < 14 * 3), so to commit any block would
require colloboration between the unlocked and locked nodes. For any
block announced, the locked nodes would require that the sealer of the
block was the original sealer. The unlocked nodes would require that the
sealer of the block is the sender. The stall came from the passage of 20
rounds, as the only acceptable proposer to both factions would be 315
again[2]. In general, progress is not guaranteed as the validator which
nodes have locked on need not be one of the locked nodes.

There are two possible corrections. One is for the reannouncers of a
locked block to unseal and reseal it with their own private key, and for
validators to compare lock replacements on a hash which ignores the
proposer seal. Two is for validators to relax the seal author checking:
instead of comparing against the believed signer (which can be a false
belief if there is disagreement about a lock), just check that the seal
author is any validator.

[0] `for dir in */; do echo $dir; grep "showctx/view.*sequence = 95" $dir/strato-sequencer | wc -l; done`
[1] `grep "21:34:06.8.*showctx/mLockedSender" */strato-sequencer`
[2] `grep "showctx/view.*sequence = 95" strato-sequencer | cut -d \| -f
3,4 | uniq -c`