Provide two streams of faucet nonces
The low stream is the nonce read from postgres, which is will
typically fail when concurrent faucets are being executed. The
high stream is an atomic counter for this node. In practice, I've
found the optimistic high stream very likely to succeed (on singlenode).
In multinode, both the low nonce and high nonce could be stolen by
another node but since there is still a continous span of nonces
it does not stop nonces permanently