Revert the ordering of detailsForRow
I had misunderstood the original implementation, and thought that
if the source is present, it should be compiled and inserted. What
happens instead with the current ordering is that multiple ActionDatas
share a metadata, and so if contract `Y` calls `new X()` both of them
are labeled as `Y`s. In hindsight, it should have been obvious that
it should have been structurally similar to `getCachedSolidVMDetails`.

This change includes one test that actually discriminates on this
problem (in slipstream.test.js) and one test from a dead end of
my investigations (ProcessorSpec.hs).