Use a reverse post order traversal to limit copies
This algorithm does two passes over the RLP. The first
calculates the length, which a buffer is allocated for.
The second does a reverse post order tree traversal (children first,
from right to left) and decrements the buffer point by the width of
the write for each child. The change in pointer position is then
used to calculate the total length of a node and written into the RLP
header.