Import/export improvements
Killing qs savekafka was difficult to do reliably in a script, because
sometimes it would be unresponsive (e.g. in docker without -i) and a
SIGKILL wouldn't flush pending writes. It seems safer to just read
all available messages instead of subscribing to future messages on
the stream.

loadkafka took hours on the 1.7GB of the SRR node in prod. Changing it
to a 1MB per message bundle improved throughput dramatically to 316s for upload.