Remember more round changes
Consider a 3 node network that is only partially connected, and then
gets reconnected when 1 node is ahead of the other two that are
on the same round.

In the current implementation, this network will be stuck forever.
However, its likely that the behind nodes have seen all the round
changes from the ahead node because it needed quorum to transition
ahead. When the behind nodes reconnect, they can use the lingering
round change from the ahead node to synchronize themselves and
step forward to the ahead node.