Fix the CertificateRegistry Contract
The root cert cannot be posted in the constructor of the
CertificateRegistry contract. If the root cert is posted in the
CR dapp the CR will be prefixed by the organization of the
rootCert (ex. BlockApps-CertificateRegistry).

To work around this, I made an initalization function. Before
anything can be registered, the contract must be initalized.