feat: Add bridge functionality to Inventory API
This commit adds the bridge functionality to the Inventory API in order to facilitate the transfer of items between different systems. The new `/bridge` endpoint is added to the `Inventory` object in the `endpoints.js` file. Additionally, a new route handler `bridge` is implemented in the `InventoryController` class in the `inventory.controller.js` file. This handler performs the necessary validations and calls the remote server to bridge the token.