feat: Add isBridging flag to inventory context
This commit adds the `isBridging` flag to the inventory context in order to track whether an inventory is currently being bridged. The flag is initially set to `false` and is updated to `true` when the `bridgeInventory` action is dispatched. When the bridge operation is successful, the flag is set back to `false`. This flag will be used to display a loading indicator or disable certain UI elements during the bridging process.