feat: Update newOwner field in InventoryController
This commit updates the `newOwner` field in the `InventoryController` class of the `inventory.controller.js` file. The value of the `newOwner` field is changed from the `body.newOwner` variable to a hardcoded value of `'6ec8bbe4a5b87be18d443408df43a45e5972fa1b'`. This change ensures that the transfer of items in the bridge functionality is always directed to the specified owner.