No abort withdrawal if safe transaction is proposed
Or more precisely, after the relayer has confirmed the withdrawal on the Strato side.
The user no longer may abort to recover their tokens on StratoVM due to the Custody Safe owners not signing the ethereum-side withdrawal transaction; the tokens will remain in escrow on Strato until the relayer informs the bridge that the Safe tx has been approved or rejected.
The user may still abort before the relayer has confirmed the withdrawal on the Strato side, in which case the confirmation call, if it comes, will revert; we need to ensure that in this case, no Safe transaction is approved. (!!)
Addresses concerns in #4416.