Introduce waitForBalanceUpdate that waits for the balance update in the indexer
Both lending.service and safety.service query Cirrus immediately after
transaction was executed on chain.

This means that in some rare conditions, the indexer is not updated on time and
the resulted lpTokenAfter is the same as lpTokenBefore, resulting in 0 tokens to
be staked for rewards.

The waitForBalanceUpdate is the simplest way to deal with this race condition.